Move QueueClientServiceExtensions
